requirements.txt
attrs==21.2.0
autopep8==1.5.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.3
chrome-gnome-shell==0.0.0
click==8.0.1
CoinMarketCapAPI==0.5
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
entrypoints==0.3
eventlet==0.31.1
Flask==2.0.1
Flask-MQTT==1.1.1
greenlet==1.1.0
hidpidaemon==18.4.6
httplib2==0.14.0
idna==2.8
iota-client-python==0.2.0a3
iota-wallet-python-binding==0.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==2.0.1
netifaces==0.10.4
oauthlib==3.1.0
paho-mqtt==1.5.1
pop-transition==1.1.2
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
pydbus==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt===2.1.2pop0-1587756471-20.04-cd2988e
python-coinmarketcap==0.2
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.4.1
ratelimit==2.2.1
repoman==1.2.2
requests==2.26.0
requests-cache==0.7.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
sessioninstaller==0.0.0
simplejson==3.16.0
six==1.14.0
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
url-normalize==1.4.3
urllib3==1.26.6
vboxapi==1.0
wadllib==1.3.3
Werkzeug==2.0.1
xkit==0.0.0
